Add common.h to avoid code duplication.
The implementations of these functions were exactly identical.
