+++ ryu/d2s.c
@@ -53,23 +53,8 @@
 #define HAS_64_BIT_INTRINSICS
 #endif
 
+#include "ryu/common.h"
 #include "ryu/d2s.h"
-
-// Returns floor(log_10(2^e)).
-static inline int32_t log10Pow2(const int32_t e) {
-  // This function has only been tested for 0 <= e <= 1500.
-  assert(e >= 0);
-  assert(e <= 1500);
-  return (int32_t) ((((uint32_t) e) * 78913) >> 18);
-}
-
-// Returns floor(log_10(5^e)).
-static inline int32_t log10Pow5(const int32_t e) {
-  // This function has only been tested for 0 <= e <= 1500.
-  assert(e >= 0);
-  assert(e <= 1500);
-  return (int32_t) ((((uint32_t) e) * 732923) >> 20);
-}
 
 static inline int32_t pow5Factor(uint64_t value) {
   for (int32_t count = 0; value > 0; ++count) {
@@ -302,7 +287,7 @@ void d2s_buffered(double f, char* result) {
     // This expression is slightly faster than max(0, log10Pow2(e2) - 1).
     const int32_t q = log10Pow2(e2) - (e2 > 3);
     e10 = q;
-    const int32_t k = DOUBLE_POW5_INV_BITCOUNT + double_pow5bits(q) - 1;
+    const int32_t k = DOUBLE_POW5_INV_BITCOUNT + pow5bits(q) - 1;
     const int32_t i = -e2 + q + k;
 #if defined(RYU_OPTIMIZE_SIZE)
     uint64_t pow5[2];
@@ -336,7 +321,7 @@ void d2s_buffered(double f, char* result) {
     const int32_t q = log10Pow5(-e2) - (-e2 > 1);
     e10 = q + e2;
     const int32_t i = -e2 - q;
-    const int32_t k = double_pow5bits(i) - DOUBLE_POW5_BITCOUNT;
+    const int32_t k = pow5bits(i) - DOUBLE_POW5_BITCOUNT;
     const int32_t j = q - k;
 #if defined(RYU_OPTIMIZE_SIZE)
     uint64_t pow5[2];
